chore: move tests with aws calls to separate files
