+++ internal/pkg/deploy/cloudformation/stack/scheduled_job_integration_test.go
@@ -12,9 +12,6 @@ import (
 
 	"gopkg.in/yaml.v3"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/cloudformation"
-	"github.com/aws/copilot-cli/internal/pkg/aws/sessions"
 	"github.com/aws/copilot-cli/internal/pkg/deploy/cloudformation/stack"
 
 	"github.com/aws/copilot-cli/internal/pkg/manifest"
@@ -39,17 +36,6 @@ func TestScheduledJob_Template(t *testing.T) {
 
 	tpl, err := serializer.Template()
 	require.NoError(t, err, "template should render")
-
-	sess, err := sessions.NewProvider().Default()
-	require.NoError(t, err)
-	cfn := cloudformation.New(sess)
-
-	t.Run("CF template should be valid", func(t *testing.T) {
-		_, err := cfn.ValidateTemplate(&cloudformation.ValidateTemplateInput{
-			TemplateBody: aws.String(tpl),
-		})
-		require.NoError(t, err)
-	})
 
 	t.Run("CF Template should be equal", func(t *testing.T) {
 		actualBytes := []byte(tpl)
